extracted geo location button into its own component
